% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconbio.R
\name{cv.pr.w}
\alias{cv.pr.w}
\title{Pseudo removed leave out cross validation}
\usage{
cv.pr.w(
  modern_taxa,
  modern_climate,
  nPLS = 5,
  trainfun,
  predictfun,
  pseduo,
  usefx = FALSE,
  fx = NA
)
}
\arguments{
\item{modern_taxa}{}

\item{modern_climate}{}

\item{nPLS}{}

\item{trainfun}{}

\item{predictfun}{}

\item{pseduo}{}

\item{usefx}{}

\item{fx}{}
}
\value{

}
\description{
Pseudo removed leave out cross validation
}
