% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxTWAPLS.R
\docType{package}
\name{fxTWAPLS-package}
\alias{fxTWAPLS}
\alias{fxTWAPLS-package}
\title{fxTWAPLS: An Improved Version of WA-PLS}
\description{
The goal of this package is to provide an improved version of WA-PLS (Weighted Averaging Partial Least Squares) by including the tolerances of taxa and the frequency of the sampled climate variable. This package also provides a way of leave-out cross-validation that removes both the test site and sites that are both geographically close and climatically close for each cycle, to avoid the risk of pseudo-replication.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/special-uor/fxTWAPLS/}
  \item \url{https://special-uor.github.io/fxTWAPLS/}
  \item \url{https://research.reading.ac.uk/palaeoclimate/}
  \item Report bugs at \url{https://github.com/special-uor/fxTWAPLS/issues/}
}

}
\author{
\strong{Maintainer}: Roberto Villegas-Diaz \email{r.villegas-diaz@outlook.com} (\href{https://orcid.org/0000-0001-5036-8661}{ORCID})

Authors:
\itemize{
  \item Mengmeng Liu \email{m.liu18@imperial.ac.uk} (\href{https://orcid.org/0000-0001-6250-0148}{ORCID})
  \item Iain Colin Prentice \email{c.prentice@imperial.ac.uk} (\href{https://orcid.org/0000-0002-1296-6764}{ORCID})
  \item Cajo J. F. ter Braak \email{cajo.terbraak@wur.nl} (\href{https://orcid.org/0000-0002-0414-8745}{ORCID})
  \item Sandy P. Harrison \email{s.p.harrison@reading.ac.uk} (\href{https://orcid.org/0000-0001-5687-1903}{ORCID})
}

Other contributors:
\itemize{
  \item SPECIAL Research Group @ University of Reading [copyright holder]
}

}
\keyword{internal}
