% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxTWAPLS.R
\name{get_distance}
\alias{get_distance}
\title{Get the distance between points}
\usage{
get_distance(point)
}
\arguments{
\item{point}{each row represents a sampling site, the first column is
longitude and the second column is latitude, both in decimal format}
}
\value{
distance matrix, the value at the ith row, means the distance between
the ith sampling site and the whole sampling sites
}
\description{
Get the distance between points
}
