% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxTWAPLS.R
\name{get_pseduo}
\alias{get_pseduo}
\title{Leave out cross validation with pseudo (geographically and climatically close)
sites removed from the training set.
Get the ones which are both geographically and climatically close, and could
therefore results in pseudo-replication.}
\usage{
get_pseduo(dist, x, cpus = 4)
}
\arguments{
\item{dist}{TODO}

\item{x}{TODO}

\item{cpus}{TODO}
}
\value{
TODO
}
\description{
Leave out cross validation with pseudo (geographically and climatically close)
sites removed from the training set.
Get the ones which are both geographically and climatically close, and could
therefore results in pseudo-replication.
}
