% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxTWAPLS.R
\name{get_pseudo}
\alias{get_pseudo}
\title{Leave out cross validation with pseudo (geographically and climatically close)
sites removed from the training set.
Get the ones which are both geographically and climatically close, and could
therefore results in pseudo-replication.}
\usage{
get_pseudo(dist, x, cpus = 4)
}
\arguments{
\item{dist}{distance matrix which contains the distance from other sites.}

\item{x}{the modern climate values}

\item{cpus}{number of CPUs for simultaneous iterations to execute, check
\code{parallel::detectCores()} for available CPUs on your machine.}
}
\value{
the geographically and climatically close sites to each test site.
}
\description{
Leave out cross validation with pseudo (geographically and climatically close)
sites removed from the training set.
Get the ones which are both geographically and climatically close, and could
therefore results in pseudo-replication.
}
\examples{
\dontrun{
    # Load modern pollen data
    modern_pollen <- read.csv(system.file("extdata", 
                                          "Modern_Pollen_gdd_alpha_Tmin.csv", 
                                          package = "fxTWAPLS", 
                                          mustWork = TRUE))
    point <- modern_pollen[, c("Long", "Lat")]
    dist <- get_distance(point, cpus = 1)
    pseudo_Tmin <- fxTWAPLS::get_pseudo(dist, 
                                        modern_pollen$Tmin, 
                                        cpus = 1)
}
}
\seealso{
\code{\link{get_distance}}
}
