% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modern_pollen}
\alias{modern_pollen}
\title{Modern pollen dataset}
\format{
A data frame with 6458 rows and 204 variables:
\itemize{
\item \strong{\code{Entity.name}:} ...
\item \strong{\code{Lat}:} ...
\item \strong{\code{Long}:} ...
\item \strong{\code{Elv}:} ...
\item \strong{\code{gdd}:} ...
\item \strong{\code{MI}:} ...
\item \strong{\code{Tmin}:} ...
\item \strong{\code{alpha}:} ...
\item \strong{\code{Hill_N1}:} Diversity index.
\item And the following specimens: \strong{\code{Abies}}, \strong{\code{Acer}}, \strong{\code{Aconitum}}, \strong{\code{Adonis}}, \strong{\code{Aesculus}}, \strong{\code{Alnus}}, \strong{\code{Alnus.alnobetula}}, \strong{\code{Amaranthaceae}}, \strong{\code{Amaryllidaceae}}, \strong{\code{Amygdaloideae}}, \strong{\code{Andromeda}}, \strong{\code{Apiaceae}}, \strong{\code{Aquilegia}}, \strong{\code{Arbutus}}, \strong{\code{Arctostaphylos}}, \strong{\code{Argania}}, \strong{\code{Artemisia}}, \strong{\code{Asparagaceae}}, \strong{\code{Asteraceae}}, \strong{\code{Asteraceae..Liguliflorae.}}, \strong{\code{Asteroideae}}, \strong{\code{Astragalus}}, \strong{\code{Berberis}}, \strong{\code{Betula}}, \strong{\code{Betula..Chamaebetula.}}, \strong{\code{Boraginaceae}}, \strong{\code{Brassicaceae}}, \strong{\code{Bruckenthalia}}, \strong{\code{Buxus}}, \strong{\code{Calluna}}, \strong{\code{Campanulaceae}}, \strong{\code{Caprifoliaceae}}, \strong{\code{Caragana}}, \strong{\code{Carduoideae}}, \strong{\code{Carpinus.betulus}}, \strong{\code{Carpinus.orientalis.Ostrya}}, \strong{\code{Caryophyllaceae}}, \strong{\code{Cassiope}}, \strong{\code{Castanea}}, \strong{\code{Cedrus}}, \strong{\code{Celastraceae}}, \strong{\code{Celtis}}, \strong{\code{Ceratonia}}, \strong{\code{Cichorioideae}}, \strong{\code{Cistaceae}}, \strong{\code{Cistus}}, \strong{\code{Clematis}}, \strong{\code{Colchicaceae}}, \strong{\code{Convolvulaceae}}, \strong{\code{Cornus}}, \strong{\code{Corylus}}, \strong{\code{Cotinus}}, \strong{\code{Crassulaceae}}, \strong{\code{Crataegus}}, \strong{\code{Cupressaceae}}, \strong{\code{Cyperaceae}}, \strong{\code{Cytinaceae}}, \strong{\code{Daphne}}, \strong{\code{Delphinium}}, \strong{\code{Dennstaedtiaceae}}, \strong{\code{Diapensia}}, \strong{\code{Dryas}}, \strong{\code{Empetrum}}, \strong{\code{Ephedra}}, \strong{\code{Equisetum}}, \strong{\code{Erica}}, \strong{\code{Ericaceae}}, \strong{\code{Euonymus}}, \strong{\code{Euphorbiaceae}}, \strong{\code{Fabaceae}}, \strong{\code{Fabaceae..herbs.}}, \strong{\code{Fagus}}, \strong{\code{Frangula}}, \strong{\code{Fraxinus}}, \strong{\code{Genisteae}}, \strong{\code{Gentianaceae}}, \strong{\code{Geraniaceae}}, \strong{\code{Halimium}}, \strong{\code{Hedera}}, \strong{\code{Helianthemum}}, \strong{\code{Helleborus}}, \strong{\code{Hippophae}}, \strong{\code{Huperzia}}, \strong{\code{Hypericaceae}}, \strong{\code{Ilex}}, \strong{\code{Impatiens}}, \strong{\code{Iridaceae}}, \strong{\code{Jasminum}}, \strong{\code{Juglans}}, \strong{\code{Juncaceae}}, \strong{\code{Kalmia}}, \strong{\code{Lamiaceae}}, \strong{\code{Larix}}, \strong{\code{Lavandula}}, \strong{\code{Ledum}}, \strong{\code{Ligustrum}}, \strong{\code{Liliaceae}}, \strong{\code{Linaceae}}, \strong{\code{Linnaea}}, \strong{\code{Linum}}, \strong{\code{Lonicera}}, \strong{\code{Lycopodium}}, \strong{\code{Lysimachia}}, \strong{\code{Lythraceae}}, \strong{\code{Malus}}, \strong{\code{Malvaceae}}, \strong{\code{Melanthiaceae}}, \strong{\code{Mercurialis}}, \strong{\code{Montiaceae}}, \strong{\code{Moraceae}}, \strong{\code{Myrica}}, \strong{\code{Myrtaceae}}, \strong{\code{Nartheciaceae}}, \strong{\code{Nerium}}, \strong{\code{Nitrariaceae}}, \strong{\code{Olea}}, \strong{\code{Oleaceae}}, \strong{\code{Onagraceae}}, \strong{\code{Ononis}}, \strong{\code{Ophioglossaceae}}, \strong{\code{Orobanchaceae}}, \strong{\code{Osmundaceae}}, \strong{\code{Oxalidaceae}}, \strong{\code{Oxyria.Rumex}}, \strong{\code{Paliurus}}, \strong{\code{Papaveraceae}}, \strong{\code{Parrotia}}, \strong{\code{Phillyrea}}, \strong{\code{Picea}}, \strong{\code{Picea.orientalis}}, \strong{\code{Pinus..diploxylon.}}, \strong{\code{Pinus..haploxylon.}}, \strong{\code{Pistacia}}, \strong{\code{Plantaginaceae}}, \strong{\code{Platanus}}, \strong{\code{Plumbaginaceae}}, \strong{\code{Poaceae}}, \strong{\code{Polemoniaceae}}, \strong{\code{Polygalaceae}}, \strong{\code{Polygonaceae}}, \strong{\code{Polygonum}}, \strong{\code{Polypodiales}}, \strong{\code{Populus}}, \strong{\code{Portulacaceae}}, \strong{\code{Potentilla}}, \strong{\code{Primulaceae}}, \strong{\code{Prunus}}, \strong{\code{Pteridaceae}}, \strong{\code{Pterocarya}}, \strong{\code{Quercus.deciduous}}, \strong{\code{Quercus.evergreen}}, \strong{\code{Quercus.intermediate}}, \strong{\code{Ranunculaceae}}, \strong{\code{Ranunculus}}, \strong{\code{Resedaceae}}, \strong{\code{Rhamnaceae}}, \strong{\code{Rhamnus}}, \strong{\code{Rhododendron}}, \strong{\code{Rhus}}, \strong{\code{Ribes}}, \strong{\code{Rosaceae}}, \strong{\code{Rosmarinus}}, \strong{\code{Rubiaceae}}, \strong{\code{Rubus}}, \strong{\code{Ruscus}}, \strong{\code{Salix}}, \strong{\code{Salvia}}, \strong{\code{Sambucus}}, \strong{\code{Sanguisorba.group}}, \strong{\code{Santalaceae}}, \strong{\code{Saxifragaceae}}, \strong{\code{Scrophulariaceae}}, \strong{\code{Solanaceae}}, \strong{\code{Sorbus}}, \strong{\code{Styrax}}, \strong{\code{Tamarix}}, \strong{\code{Taxus}}, \strong{\code{Teucrium}}, \strong{\code{Thalictrum}}, \strong{\code{Thymelaeaceae}}, \strong{\code{Tilia}}, \strong{\code{Tofieldia}}, \strong{\code{Trollius}}, \strong{\code{Ulmus}}, \strong{\code{Ulmus.Zelkova}}, \strong{\code{Urticaceae}}, \strong{\code{Vaccinium}}, \strong{\code{Valerianaceae}}, \strong{\code{Verbenaceae}}, \strong{\code{Viburnum}}, \strong{\code{Violaceae}}, \strong{\code{Viscum}}, \strong{\code{Vitex}}, \strong{\code{Ziziphus}}, \strong{\code{Zygophyllaceae}}.
}
}
\source{
\url{https://research.reading.ac.uk/palaeoclimate/}
}
\usage{
data(modern_pollen)
}
\description{
A dataset containing information of ...
}
\author{
Given Surname \email{given.surname@email}
}
\keyword{datasets}
