% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxTWAPLS.R
\name{plot.residuals}
\alias{plot.residuals}
\title{Plot the residuals of the training results}
\usage{
\method{plot}{residuals}(train_output, col)
}
\arguments{
\item{train_output}{Training output, can be the output of WAPLS, WAPLS with
fx correction, TWAPLS, or TWAPLS with fx correction}

\item{col}{choose which column of the fitted value to plot, in other words,
how many number of components you want to use}
}
\description{
Plot the residuals of the training results
}
