% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxTWAPLS.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot the residuals of the training results}
\usage{
plot_residuals(train_output, col)
}
\arguments{
\item{train_output}{Training output, can be the output of WAPLS, WAPLS with
fx correction, TWAPLS, or TWAPLS with fx correction}

\item{col}{choose which column of the fitted value to plot, in other words,
how many number of components you want to use}
}
\description{
Plot the residuals of the training results
}
\seealso{
\code{\link{TWAPLS.w}} and \code{\link{WAPLS.w}}
}
