% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxTWAPLS.R
\name{rand.t.test.w}
\alias{rand.t.test.w}
\title{Random t-test}
\usage{
rand.t.test.w(cvoutput, n.perm = 999)
}
\arguments{
\item{cvoutput}{cross-validation output either from \code{\link{cv.w}} or
\code{\link{cv.pr.w}}}

\item{n.perm}{the number of permutation times to get the p value, which
assesses whether using the current number of components is significantly
different from using one less.}
}
\value{
a matrix of the statistics of the cross-validation results
}
\description{
Random t-test
}
