% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconbio.R
\name{sse.sample}
\alias{sse.sample}
\title{Calculate Sample Specific Errors}
\usage{
sse.sample(
  modern_taxa,
  modern_climate,
  fossil_taxa,
  trainfun,
  predictfun,
  nboot,
  nPLS,
  nsig,
  usefx,
  fx
)
}
\arguments{
\item{modern_taxa}{}

\item{modern_climate}{}

\item{fossil_taxa}{}

\item{trainfun}{}

\item{predictfun}{}

\item{nboot}{}

\item{nPLS}{}

\item{nsig}{}

\item{usefx}{}

\item{fx}{}
}
\value{

}
\description{
Calculate Sample Specific Errors
}
