% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxTWAPLS.R
\name{TWAPLS.predict.w}
\alias{TWAPLS.predict.w}
\title{TWAPLS predict function}
\usage{
TWAPLS.predict.w(TWAPLSoutput, fossil_taxa)
}
\arguments{
\item{TWAPLSoutput}{the output of the \code{\link{TWAPLS.w}} training
function, either with or without fx correction}

\item{fossil_taxa}{fossil taxa abundance data to reconstruct past climates,
each row represents a site to be reconstructed, each column represents
a taxon.}
}
\value{
a list of the reconstruction results. fit is the fitted value.
}
\description{
TWAPLS predict function
}
