% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconbio.R
\name{TWAPLS.w}
\alias{TWAPLS.w}
\title{TWALPS training function fit}
\usage{
TWAPLS.w(modern_taxa, modern_climate, nPLS = 5, usefx = FALSE, fx = NA)
}
\arguments{
\item{modern_taxa}{}

\item{modern_climate}{}

\item{nPLS}{}

\item{usefx}{}

\item{fx}{}
}
\value{

}
\description{
TWALPS training function fit
}
