% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconbio.R
\name{cv.pr.w}
\alias{cv.pr.w}
\title{Pseudo removed leave out cross validation}
\usage{
cv.pr.w(
  modern_taxa,
  modern_climate,
  nPLS = 5,
  trainfun,
  predictfun,
  pseduo,
  usefx = FALSE,
  fx = NA,
  cpus = 4
)
}
\arguments{
\item{modern_taxa}{TODO}

\item{modern_climate}{TODO}

\item{nPLS}{TODO}

\item{trainfun}{TODO}

\item{predictfun}{TODO}

\item{pseduo}{TODO}

\item{usefx}{TODO}

\item{fx}{TODO}

\item{cpus}{TODO}
}
\value{
TODO
}
\description{
Pseudo removed leave out cross validation
}
