% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxTWAPLS.R
\name{cv.w}
\alias{cv.w}
\title{Leave-one-out cross-validation}
\usage{
cv.w(
  modern_taxa,
  modern_climate,
  nPLS = 5,
  trainfun,
  predictfun,
  usefx = FALSE,
  fx_method = "bin",
  bin = NA,
  cpus = 4,
  test_mode = FALSE,
  test_it = 5
)
}
\arguments{
\item{modern_taxa}{The modern taxa abundance data, each row represents a
sampling site, each column represents a taxon.}

\item{modern_climate}{The modern climate value at each sampling site.}

\item{nPLS}{The number of components to be extracted.}

\item{trainfun}{Training function you want to use, either
\code{\link{WAPLS.w}} or \code{\link{TWAPLS.w}}.}

\item{predictfun}{Predict function you want to use: if \code{trainfun} is
\code{\link{WAPLS.w}}, then this should be \code{\link{WAPLS.predict.w}};
if \code{trainfun} is \code{\link{TWAPLS.w}}, then this should be
\code{\link{TWAPLS.predict.w}}.}

\item{usefx}{Boolean flag on whether or not use \code{fx} correction.}

\item{fx_method}{Binned or p-spline smoothed \code{fx} correction: if
\code{usefx = FALSE}, this should be \code{NA}; otherwise,
\code{\link{fx}} function will be used when choosing "bin";
\code{\link{fx_pspline}} function will be used when choosing "pspline".}

\item{bin}{Binwidth to get fx, needed for both binned and p-splined method.
if \code{usefx = FALSE}, this should be \code{NA};}

\item{cpus}{Number of CPUs for simultaneous iterations to execute, check
\code{parallel::detectCores()} for available CPUs on your machine.}

\item{test_mode}{boolean flag to execute the function with a limited number
of iterations, \code{test_it}, for testing purposes only.}

\item{test_it}{number of iterations to use in the test mode.}
}
\value{
leave-one-out cross validation results
}
\description{
Leave-one-out cross-validation as
\code{rioja} (\url{https://cran.r-project.org/package=rioja}).
}
\examples{
\dontrun{
# Load modern pollen data
modern_pollen <- read.csv("/path/to/modern_pollen.csv")
                                      
# Extract taxa
taxaColMin <- which(colnames(modern_pollen) == "taxa0")
taxaColMax <- which(colnames(modern_pollen) == "taxaN")
taxa <- modern_pollen[, taxaColMin:taxaColMax]

## LOOCV
test_mode <- TRUE # It should be set to FALSE before running
### without fx
cv_t_Tmin <- fxTWAPLS::cv.w(taxa,
                            modern_pollen$Tmin,
                            nPLS = 5,
                            fxTWAPLS::TWAPLS.w,
                            fxTWAPLS::TWAPLS.predict.w,
                            cpus = 2, # Remove the following line
                            test_mode = test_mode)
### with fx
cv_tf_Tmin <- fxTWAPLS::cv.w(taxa,
                             modern_pollen$Tmin,
                             nPLS = 5,
                             fxTWAPLS::TWAPLS.w,
                             fxTWAPLS::TWAPLS.predict.w,
                             usefx = TRUE,
                             fx_method = "bin",
                             bin = 0.02,
                             cpus = 2, # Remove the following line
                             test_mode = test_mode)  

# Run with progress bar
`\%>\%` <- magrittr::`\%>\%`
### without fx
cv_t_Tmin <- fxTWAPLS::cv.w(taxa,
                            modern_pollen$Tmin,
                            nPLS = 5,
                            fxTWAPLS::TWAPLS.w,
                            fxTWAPLS::TWAPLS.predict.w,
                            cpus = 2, # Remove the following line
                            test_mode = test_mode) \%>\% fxTWAPLS::pb()
### with fx
cv_tf_Tmin <- fxTWAPLS::cv.w(taxa,
                             modern_pollen$Tmin,
                             nPLS = 5,
                             fxTWAPLS::TWAPLS.w,
                             fxTWAPLS::TWAPLS.predict.w,
                             usefx = TRUE,
                             fx_method = "bin",
                             bin = 0.02,
                             cpus = 2, # Remove the following line
                             test_mode = test_mode) \%>\% fxTWAPLS::pb()
}
}
\seealso{
\code{\link{fx}}, \code{\link{TWAPLS.w}},
\code{\link{TWAPLS.predict.w}}, \code{\link{WAPLS.w}}, and
\code{\link{WAPLS.predict.w}}
}
