% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconbio.R
\name{cv.w}
\alias{cv.w}
\title{Leave one out cross validation as rioja}
\usage{
cv.w(
  modern_taxa,
  modern_climate,
  nPLS = 5,
  trainfun,
  predictfun,
  usefx = FALSE,
  fx = NA
)
}
\arguments{
\item{modern_taxa}{}

\item{modern_climate}{}

\item{nPLS}{}

\item{trainfun}{}

\item{predictfun}{}

\item{usefx}{}

\item{fx}{}
}
\value{

}
\description{
Leave one out cross validation as rioja
}
