% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbind_pb}
\alias{rbind_pb}
\title{rbind combine with progress bar, to be used in combination with foreach}
\usage{
rbind_pb(iterator)
}
\arguments{
\item{iterator}{TODO}
}
\description{
rbind combine with progress bar, to be used in combination with foreach
}
