% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxTWAPLS.R
\name{sse.sample}
\alias{sse.sample}
\title{Calculate Sample Specific Errors}
\usage{
sse.sample(
  modern_taxa,
  modern_climate,
  fossil_taxa,
  trainfun,
  predictfun,
  nboot,
  nPLS,
  nsig,
  usefx,
  fx
)
}
\arguments{
\item{modern_taxa}{the modern taxa abundance data, each row represents a
sampling site, each column represents a taxon.}

\item{modern_climate}{the modern climate value at each sampling site}

\item{fossil_taxa}{fossil taxa abundance data to reconstruct past climates,
each row represents a site to be reconstructed, each column represents a
taxon.}

\item{trainfun}{training function you want to use, either
\code{\link{WAPLS.w}} or \code{\link{TWAPLS.w}}}

\item{predictfun}{predict function you want to use: if \code{trainfun} is
\code{\link{WAPLS.w}}, then this should be \code{\link{WAPLS.predict.w}};
if \code{trainfun} is \code{\link{TWAPLS.w}}, then this should be
\code{\link{TWAPLS.predict.w}}}

\item{nboot}{the number of bootstrap cycles you want to use}

\item{nPLS}{the number of components to be extracted}

\item{nsig}{the significant number of components to use to reconstruct past
climates, this can be obtained from the cross-validation results.}

\item{usefx}{boolean flag on whether or not use fx correction.}

\item{fx}{the frequency of the climate value for fx correction: if
\code{usefx} is FALSE, this should be \code{NA}; otherwise, this should
be obtained from the \code{\link{fx}} function.}
}
\value{
the bootstrapped standard error for each site
}
\description{
Calculate Sample Specific Errors
}
