% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconbio.R
\name{sse.sample}
\alias{sse.sample}
\title{Calculate Sample Specific Errors}
\usage{
sse.sample(
  modern_taxa,
  modern_climate,
  fossil_taxa,
  trainfun,
  predictfun,
  nboot,
  nPLS,
  nsig,
  usefx,
  fx
)
}
\arguments{
\item{modern_taxa}{TODO}

\item{modern_climate}{TODO}

\item{fossil_taxa}{TODO}

\item{trainfun}{TODO}

\item{predictfun}{TODO}

\item{nboot}{TODO}

\item{nPLS}{TODO}

\item{nsig}{TODO}

\item{usefx}{TODO}

\item{fx}{TODO}
}
\value{
TODO
}
\description{
Calculate Sample Specific Errors
}
