% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconbio.R
\name{TWAPLS.predict.w}
\alias{TWAPLS.predict.w}
\title{TWAPLS predict function}
\usage{
TWAPLS.predict.w(TWAPLSoutput, fossil_taxa)
}
\arguments{
\item{TWAPLSoutput}{}

\item{fossil_taxa}{}
}
\value{

}
\description{
TWAPLS predict function
}
