% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxTWAPLS.R
\name{TWAPLS.predict.w}
\alias{TWAPLS.predict.w}
\title{TWAPLS predict function}
\usage{
TWAPLS.predict.w(TWAPLSoutput, fossil_taxa)
}
\arguments{
\item{TWAPLSoutput}{TODO}

\item{fossil_taxa}{TODO}
}
\value{
TODO
}
\description{
TWAPLS predict function
}
