% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxTWAPLS.R
\name{TWAPLS.w}
\alias{TWAPLS.w}
\title{TWA-PLS training function}
\usage{
TWAPLS.w(modern_taxa, modern_climate, nPLS = 5, usefx = FALSE, fx = NA)
}
\arguments{
\item{modern_taxa}{The modern taxa abundance data, each row represents a
sampling site, each column represents a taxon.}

\item{modern_climate}{The modern climate value at each sampling site.}

\item{nPLS}{The number of components to be extracted.}

\item{usefx}{Boolean flag on whether or not use \code{fx} correction.}

\item{fx}{The frequency of the climate value for \code{fx} correction: if
\code{usefx = FALSE}, this should be \code{NA}; otherwise, this should
be obtained from the \code{\link{fx}} function.}
}
\value{
A list of the training results, which will be used by the predict
function. Each element in the list is described below:
\describe{
\item{\code{fit}}{the fitted values using each number of components.}
\item{\code{x}}{the observed modern climate values.}
\item{\code{taxon_name}}{the name of each taxon.}
\item{\code{optimum}}{the updated taxon optimum}
\item{\code{comp}}{each component extracted (will be used in step 7
regression).}
\item{\code{u}}{taxon optimum for each component (step 2).}
\item{\code{t}}{taxon tolerance for each component (step 2).}
\item{\code{z}}{a parameter used in standardization for each component
(step 5).}
\item{\code{s}}{a parameter used in standardization for each component
(step 5).}
\item{\code{orth}}{a list that stores orthogonalization parameters
(step 4).}
\item{\code{alpha}}{a list that stores regression coefficients (step 7).}
\item{\code{meanx}}{mean value of the observed modern climate values.}
\item{\code{nPLS}}{the total number of components extracted.}
}
}
\description{
TWA-PLS training function, which can perform \code{fx} correction.
}
\examples{
\dontrun{
# Load modern pollen data
modern_pollen <- read.csv("/path/to/modern_pollen.csv")
                                      
# Extract taxa
taxaColMin <- which(colnames(modern_pollen) == "taxa0")
taxaColMax <- which(colnames(modern_pollen) == "taxaN")
taxa <- modern_pollen[, taxaColMin:taxaColMax]

# Get the frequency of each climate variable fx
fx_Tmin <- fxTWAPLS::fx(modern_pollen$Tmin, bin = 0.02)
fx_gdd <- fxTWAPLS::fx(modern_pollen$gdd, bin = 20)
fx_alpha <- fxTWAPLS::fx(modern_pollen$alpha, bin = 0.002)

# MTCO
fit_t_Tmin <- fxTWAPLS::TWAPLS.w(taxa, modern_pollen$Tmin, nPLS = 5)
fit_tf_Tmin <- fxTWAPLS::TWAPLS.w(taxa, 
                                  modern_pollen$Tmin, 
                                  nPLS = 5, 
                                  usefx = TRUE, 
                                  fx = fx_Tmin)
}

}
\seealso{
\code{\link{fx}}, \code{\link{TWAPLS.predict.w}}, and
\code{\link{WAPLS.w}}
}
