% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxTWAPLS.R
\name{TWAPLS.w}
\alias{TWAPLS.w}
\title{TWALPS training function fit}
\usage{
TWAPLS.w(modern_taxa, modern_climate, nPLS = 5, usefx = FALSE, fx = NA)
}
\arguments{
\item{modern_taxa}{TODO}

\item{modern_climate}{TODO}

\item{nPLS}{TODO}

\item{usefx}{TODO}

\item{fx}{TODO}
}
\value{
TODO
}
\description{
TWALPS training function fit
}
