% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconbio.R
\name{WAPLS.predict.w}
\alias{WAPLS.predict.w}
\title{WAPLS predict function}
\usage{
WAPLS.predict.w(WAPLSoutput, fossil_taxa)
}
\arguments{
\item{WAPLSoutput}{}

\item{fossil_taxa}{}
}
\value{

}
\description{
WAPLS predict function
}
