% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxTWAPLS.R
\name{WAPLS.predict.w}
\alias{WAPLS.predict.w}
\title{WAPLS predict function}
\usage{
WAPLS.predict.w(WAPLSoutput, fossil_taxa)
}
\arguments{
\item{WAPLSoutput}{the output of the \code{\link{WAPLS.w}} training function,
either with or without fx correction}

\item{fossil_taxa}{fossil taxa abundance data to reconstruct past climates,
each row represents a site to be reconstructed, each column represents a
taxon.}
}
\value{
a list of the reconstruction results. fit is the fitted value.
}
\description{
WAPLS predict function
}
\seealso{
\code{\link{WAPLS.w}}
}
