% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconbio.R
\name{WAPLS.predict.w}
\alias{WAPLS.predict.w}
\title{WAPLS predict function}
\usage{
WAPLS.predict.w(WAPLSoutput, fossil_taxa)
}
\arguments{
\item{WAPLSoutput}{TODO}

\item{fossil_taxa}{TODO}
}
\value{
TODO
}
\description{
WAPLS predict function
}
