% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconbio.R
\name{WAPLS.w}
\alias{WAPLS.w}
\title{WALPS training function fit}
\usage{
WAPLS.w(modern_taxa, modern_climate, nPLS = 5, usefx = FALSE, fx = NA)
}
\arguments{
\item{modern_taxa}{TODO}

\item{modern_climate}{TODO}

\item{nPLS}{TODO}

\item{usefx}{TODO}

\item{fx}{TODO}
}
\value{
TODO
}
\description{
WALPS training function fit
}
